backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
boto3>=1.34.0
botocore>=1.34.0
pydantic>=2.5.2
pydantic-settings>=2.1.0
httpx>=0.25.2
apscheduler>=3.10.4
cryptography>=41.0.7
python-dateutil>=2.8.2
feedparser>=6.0.11
cachetools>=5.3.2
openpyxl>=3.1.2
